Marathon refactoring and docs
