+++ packages/visual-tests/src/TestCases/PieChart/donut-2.ts
@@ -1,5 +1,5 @@
 import { PieChart } from "@operational/visualizations"
-import { MarathonEnvironment } from "../../components/Marathon"
+import { MarathonEnvironment } from "../../Marathon"
 
 const DonutRenderer = {
   type: "donut"
