+++ packages/visual-tests/src/TestCases/PieChart/gauge-6.ts
@@ -1,5 +1,5 @@
 import { PieChart } from "@operational/visualizations"
-import { MarathonEnvironment } from "../../components/Marathon"
+import { MarathonEnvironment } from "../../Marathon"
 
 const GaugeRenderer = {
   type: "gauge",
