+++ packages/visual-tests/src/TestCases/Sunburst/case09.ts
@@ -1,5 +1,5 @@
 import { Sunburst } from "@operational/visualizations"
-import { MarathonEnvironment } from "../../components/Marathon"
+import { MarathonEnvironment } from "../../Marathon"
 
 const config = {
   maxRings: 4
